refactor: fix clippy lints and format all crates
- Auto-fix all applicable clippy suggestions per crate (collapsible_if,
  question_mark, while_let_loop, needless_map_or, redundant_closure,
  iter_cloned_collect, unnecessary_unwrap_or_else, etc.)
- Replace let/else patterns with ? operator throughout aivi-hir,
  aivi-syntax, aivi-runtime, aivi-lsp, aivi-typing, aivi-backend
- Convert loop/match to while-let in aivi-typing, aivi-hir,
  aivi-backend, aivi-lsp
- Fix boxed_local parameters in aivi-typing::source_contracts
- Replace clone-for-comparison with std::slice::from_ref
- Replace filter_map-always-Some with map in aivi-lsp::diagnostics
- Replace Iterator::last with rfind on DoubleEndedIterator
- Replace Iterator::fold with try_fold in aivi-lsp::type_annotations
- Convert match-returns-bool to matches! macro in aivi-hir::lower
- Remove redundant binding redefinition in aivi-hir::lower
- Fix loop variable indexing with enumerate in aivi-hir::lower::helpers
- Collapse identical if-else blocks in aivi-cli::manual_snippets
- Add #[allow(enum_variant_names)] for local Task enums in aivi-core
- Fix doc list overindentation in aivi-gtk::executor
- Run cargo fmt --all

+++ crates/aivi-backend/src/runtime/evaluator.rs
@@ -1200,13 +1200,13 @@ impl<'a> KernelEvaluator<'a> {
                 mut bound_arguments,
             } => {
                 bound_arguments.extend(arguments.into_iter().map(strip_signal));
-                if bound_arguments.len() < handle.field_count as usize {
+                if bound_arguments.len() < handle.field_count {
                     return Ok(RuntimeValue::Callable(RuntimeCallable::SumConstructor {
                         handle,
                         bound_arguments,
                     }));
                 }
-                let remaining = bound_arguments.split_off(handle.field_count as usize);
+                let remaining = bound_arguments.split_off(handle.field_count);
                 let value = RuntimeValue::Sum(RuntimeSumValue {
                     item: handle.item,
                     type_name: handle.type_name.clone(),
